删除用户功能优化

+++ src/views/users/User.vue
@@ -389,21 +389,38 @@ export default {
       })
     },
     handleDelete (id) {
-      deleteUser(id).then(res => {
-        console.log(res)
-        if (res.meta.status === 200) {
-          this.$message({
-            message: res.meta.msg,
-            type: 'success'
-          })
-          this.init()
-        } else {
-          this.$message({
-            message: res.meta.msg,
-            type: 'error'
-          })
-        }
+      this.$confirm('此操作将永久删除该用户, 是否继续?', '提示', {
+        confirmButtonText: '确定',
+        cancelButtonText: '取消',
+        type: 'warning'
       })
+        .then(() => {
+          deleteUser(id).then(res => {
+            console.log(res)
+            if (res.meta.status === 200) {
+              this.$message({
+                message: res.meta.msg,
+                type: 'success'
+              })
+              this.init()
+            } else {
+              this.$message({
+                message: res.meta.msg,
+                type: 'error'
+              })
+            }
+          })
+          this.$message({
+            type: 'success',
+            message: '删除成功!'
+          })
+        })
+        .catch(() => {
+          this.$message({
+            type: 'info',
+            message: '已取消删除'
+          })
+        })
     }
   }
 }
